Move RequestUtils into controllers::prelude
