+++ src/util/mod.rs
@@ -1,11 +1,10 @@
 use std::collections::HashMap;
 use std::error::Error;
-use std::io::{self, Cursor};
+use std::io::Cursor;
 use std::sync::Arc;
 
 use serde_json;
 use serde::Serialize;
-use url;
 
 use conduit::{Handler, Request, Response};
 use conduit_router::{RequestParams, RouteBuilder};
@@ -23,15 +22,6 @@ mod hasher;
 mod io_util;
 mod request_proxy;
 
-pub trait RequestUtils {
-    fn redirect(&self, url: String) -> Response;
-
-    fn json<T: Serialize>(&self, t: &T) -> Response;
-    fn query(&self) -> HashMap<String, String>;
-    fn wants_json(&self) -> bool;
-    fn pagination(&self, default: usize, max: usize) -> CargoResult<(i64, i64)>;
-}
-
 pub fn json_response<T: Serialize>(t: &T) -> Response {
     let json = serde_json::to_string(t).unwrap();
     let mut headers = HashMap::new();
@@ -44,57 +34,6 @@ pub fn json_response<T: Serialize>(t: &T) -> Response {
         status: (200, "OK"),
         headers: headers,
         body: Box::new(Cursor::new(json.into_bytes())),
-    }
-}
-
-impl<'a> RequestUtils for Request + 'a {
-    fn json<T: Serialize>(&self, t: &T) -> Response {
-        json_response(t)
-    }
-
-    fn query(&self) -> HashMap<String, String> {
-        url::form_urlencoded::parse(self.query_string().unwrap_or("").as_bytes())
-            .map(|(a, b)| (a.into_owned(), b.into_owned()))
-            .collect()
-    }
-
-    fn redirect(&self, url: String) -> Response {
-        let mut headers = HashMap::new();
-        headers.insert("Location".to_string(), vec![url.to_string()]);
-        Response {
-            status: (302, "Found"),
-            headers: headers,
-            body: Box::new(io::empty()),
-        }
-    }
-
-    fn wants_json(&self) -> bool {
-        self.headers()
-            .find("Accept")
-            .map(|accept| accept.iter().any(|s| s.contains("json")))
-            .unwrap_or(false)
-    }
-
-    fn pagination(&self, default: usize, max: usize) -> CargoResult<(i64, i64)> {
-        let query = self.query();
-        let page = query
-            .get("page")
-            .and_then(|s| s.parse::<usize>().ok())
-            .unwrap_or(1);
-        let limit = query
-            .get("per_page")
-            .and_then(|s| s.parse::<usize>().ok())
-            .unwrap_or(default);
-        if limit > max {
-            return Err(human(&format_args!(
-                "cannot request more than {} items",
-                max
-            )));
-        }
-        if page == 0 {
-            return Err(human("page indexing starts from 1, page 0 is invalid"));
-        }
-        Ok((((page - 1) * limit) as i64, limit as i64))
     }
 }
