+++ src/util/request_proxy.rs
@@ -35,7 +35,7 @@ impl<'a> Request for RequestProxy<'a> {
         self.other.virtual_root()
     }
     fn path(&self) -> &str {
-        self.path.map(|s| &*s).unwrap_or_else(|| self.other.path())
+        self.path.unwrap_or_else(|| self.other.path())
     }
     fn query_string(&self) -> Option<&str> {
         self.other.query_string()
